move get_layer_data to ultrack/utils/napari and deduplicate
The CLI already had the same multiscale-extraction pattern inline in
flow.py, link.py, export.py, and as a private function in segment.py.
Consolidate all five copies into a single get_layer_data() in
ultrack/utils/napari.py and update every call site to import from
there.
